Abstract to focus trap hook
